upgraded to typescript
